Clean up functions and add documentation
