+++ src/components/OSCALMetadata.js
@@ -75,15 +75,17 @@ export default function OSCALMetadata(props) {
             canEdit={props.isEditable}
             editedField={
               props.isEditable
-                ? [Object.keys(props.restData)[0], "metadata", "title"]
+                ? [Object.keys(props.partialRestData)[0], "metadata", "title"]
                 : null
             }
             onFieldSave={props.onFieldSave}
-            restData={
+            partialRestData={
               props.isEditable
                 ? {
-                    [Object.keys(props.restData)[0]]: {
-                      uuid: props.restData[Object.keys(props.restData)[0]].uuid,
+                    [Object.keys(props.partialRestData)[0]]: {
+                      uuid: props.partialRestData[
+                        Object.keys(props.partialRestData)[0]
+                      ].uuid,
                       metadata: {
                         title: props.metadata.title,
                       },
@@ -147,16 +149,21 @@ export default function OSCALMetadata(props) {
                   className={classes.OSCALMetadataVersion}
                   editedField={
                     props.isEditable
-                      ? [Object.keys(props.restData)[0], "metadata", "version"]
+                      ? [
+                          Object.keys(props.partialRestData)[0],
+                          "metadata",
+                          "version",
+                        ]
                       : null
                   }
                   onFieldSave={props.onFieldSave}
-                  restData={
+                  partialRestData={
                     props.isEditable
                       ? {
-                          [Object.keys(props.restData)[0]]: {
-                            uuid: props.restData[Object.keys(props.restData)[0]]
-                              .uuid,
+                          [Object.keys(props.partialRestData)[0]]: {
+                            uuid: props.partialRestData[
+                              Object.keys(props.partialRestData)[0]
+                            ].uuid,
                             metadata: {
                               version: props.metadata.version,
                             },
